backend/EPITAssoAPI/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==3.10.0
setuptools==69.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests-oauthlib==2.0.0
djangorestframework-simplejwt==5.3.1
pkce==1.0.3
drf-spectacular[sidecar]==0.27.2
channels[daphne]==4.1.0
channels-redis==4.2.0
pillow==10.4.0
django-redis==5.4.0